# src/main/java/com/nttd/banking/auth/infrastructure/adapter/in/rest/AuthApiDelegateImpl.java
package com.nttd.banking.auth.infrastructure.adapter.in.rest;

import com.nttd.banking.auth.api.ApiApiDelegate;
import com.nttd.banking.auth.application.mapper.AuthMapper;
import com.nttd.banking.auth.domain.port.in.GetJwksUseCase;
import com.nttd.banking.auth.domain.port.in.LoginUseCase;
import com.nttd.banking.auth.domain.port.in.LogoutUseCase;
import com.nttd.banking.auth.domain.port.in.RefreshTokenUseCase;
import com.nttd.banking.auth.domain.port.in.RegisterUseCase;
import com.nttd.banking.auth.domain.port.in.ValidateTokenUseCase;
import com.nttd.banking.auth.model.dto.JwkKey;
import com.nttd.banking.auth.model.dto.JwksResponse;
import com.nttd.banking.auth.model.dto.LoginRequest;
import com.nttd.banking.auth.model.dto.LoginResponse;
import com.nttd.banking.auth.model.dto.LogoutResponse;
import com.nttd.banking.auth.model.dto.RefreshTokenRequest;
import com.nttd.banking.auth.model.dto.RefreshTokenResponse;
import com.nttd.banking.auth.model.dto.RegisterRequest;
import com.nttd.banking.auth.model.dto.RegisterResponse;
import com.nttd.banking.auth.model.dto.ValidateTokenRequest;
import com.nttd.banking.auth.model.dto.ValidateTokenResponse;
import java.time.ZoneOffset;
import java.util.stream.Collectors;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.stereotype.Component;
import org.springframework.web.server.ServerWebExchange;
import reactor.core.publisher.Mono;

/**
 * Implementation of OpenAPI-generated ApiApiDelegate.
 * Only loads when not in test profile.
 */
@Component
@org.springframework.context.annotation.Profile("!test")
@RequiredArgsConstructor
@Slf4j
public class AuthApiDelegateImpl implements ApiApiDelegate {

  private final LoginUseCase loginUseCase;
  private final RegisterUseCase registerUseCase;
  private final LogoutUseCase logoutUseCase;
  private final ValidateTokenUseCase validateTokenUseCase;
  private final RefreshTokenUseCase refreshTokenUseCase;
  private final GetJwksUseCase getJwksUseCase;
  private final AuthMapper mapper;

  @Override
  public Mono<ResponseEntity<LoginResponse>> login(
      Mono<LoginRequest> loginRequest,
      ServerWebExchange exchange) {

    return loginRequest
        .flatMap(req -> loginUseCase.login(req.getUsername(), req.getPassword()))
        .map(mapper::toLoginResponse)
        .map(ResponseEntity::ok)
        .doOnSuccess(res -> log.info("User logged in successfully"));
  }

  @Override
  public Mono<ResponseEntity<RegisterResponse>> register(
      Mono<RegisterRequest> registerRequest,
      ServerWebExchange exchange) {

    return registerRequest
        .map(mapper::toRegisterRequest)
        .flatMap(registerUseCase::register)
        .map(mapper::toRegisterResponse)
        .map(res -> ResponseEntity.status(HttpStatus.CREATED).body(res))
        .doOnSuccess(res -> log.info("User registered successfully"));
  }

  @Override
  public Mono<ResponseEntity<LogoutResponse>> logout(ServerWebExchange exchange) {
    // Extract token from Authorization header
    String authHeader = exchange.getRequest().getHeaders().getFirst("Authorization");
    if (authHeader == null || !authHeader.startsWith("Bearer ")) {
      return Mono.just(ResponseEntity.status(HttpStatus.UNAUTHORIZED).build());
    }

    String token = authHeader.substring(7);

    return logoutUseCase.logout(token)
        .then(Mono.fromCallable(() -> {
          LogoutResponse response = new LogoutResponse();
          response.setMessage("Logout successful");
          return ResponseEntity.ok(response);
        }))
        .doOnSuccess(res -> log.info("User logged out successfully"));
  }

  @Override
  public Mono<ResponseEntity<ValidateTokenResponse>> validateToken(
      Mono<ValidateTokenRequest> validateTokenRequest,
      ServerWebExchange exchange) {

    return validateTokenRequest
        .flatMap(req -> validateTokenUseCase.validate(req.getToken()))
        .map(jwtToken -> {
          ValidateTokenResponse response = new ValidateTokenResponse();
          response.setValid(true);
          response.setUserId(jwtToken.getUserId());
          response.setUsername(jwtToken.getUsername());
          response.setRoles(jwtToken.getRoles());
          response.setUserType(jwtToken.getUserType());
          response.setExpiresAt(jwtToken.getExpiresAt().atOffset(ZoneOffset.UTC));
          return ResponseEntity.ok(response);
        })
        .doOnSuccess(res -> log.info("Token validated successfully"));
  }

  @Override
  public Mono<ResponseEntity<RefreshTokenResponse>> refreshToken(
      Mono<RefreshTokenRequest> refreshTokenRequest,
      ServerWebExchange exchange) {

    return refreshTokenRequest
        .flatMap(req -> refreshTokenUseCase.refresh(req.getRefreshToken()))
        .map(result -> {
          RefreshTokenResponse response = new RefreshTokenResponse();
          response.setAccessToken(result.accessToken());
          response.setTokenType("Bearer");
          response.setExpiresIn(result.expiresIn());
          return ResponseEntity.ok(response);
        })
        .doOnSuccess(res -> log.info("Token refreshed successfully"));
  }

  @Override
  public Mono<ResponseEntity<JwksResponse>> getJwks(ServerWebExchange exchange) {
    return getJwksUseCase.getJwks()
        .map(jwks -> {
          JwksResponse response = new JwksResponse();
          response.setKeys(
              jwks.getKeys().stream()
                  .map(key -> {
                    JwkKey jwkKey = new JwkKey();
                    jwkKey.setKty(key.getKty());
                    jwkKey.setAlg(key.getAlg());
                    jwkKey.setUse(key.getUse());
                    jwkKey.setKid(key.getKid());
                    jwkKey.setN(key.getN());
                    jwkKey.setE(key.getE());
                    return jwkKey;
                  })
                  .collect(Collectors.toList())
          );
          return ResponseEntity.ok(response);
        })
        .doOnSuccess(res -> log.debug("JWKS retrieved successfully"));
  }
}